ドア解錠/施錠処理のモジュール化
